Remove nginx.conf validation step
The validation step doesn't work with included config files which
reference files that use relative paths.  This means that a role down
the line could break the validation and thus the entire run by using a
relative path.  Its more easy to simply sanity check this file (since
that's the only one that was being checked by the validate step) and
be done with it.